comments added

+++ FoodDeliveryApp/src/dto/OrderDTO.java
@@ -6,6 +6,6 @@ import beans.ShoppingCart;
 public class OrderDTO {
 	public ShoppingCart cart;
 	public Address address;
-	public int points;
+	public double points;
 
 }
